requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
jupyter==1.0.0
notebook==6.5.4
pyyaml==6.0.1
xgboost==2.0.0
lightgbm==4.1.0
catboost==1.2.2 